exproted dialog to new file, added explaining headers
